AppAttach + App-Layer communication + CertStore
AppAttach procedure can be triggered by a Fetch fo an ImageStoreUsage.
The result of an AppAttach is a direct connection between the two peers
(TLS connection).

Some first attempts to use this direct connection at application-layer
level are tried out with ApplicationConnectivity class and
ImageStoreApplication (Resource name is send and received over the
direct connection)

The implementation of CertificateStore usage is in progress. Once it is
done, the Subject name in the certificates should not require a pattern.

BUGs:
- Fetching data stored at bootsrap node causes exception